@@ -4,6 +4,7 @@ import SponsorFeature
 import SwiftUI
 import VenueFeature
 
+#if !SKIP
 @main
 struct TrySwiftTokyoApp: App {
   var body: some Scene {
@@ -12,11 +13,14 @@ struct TrySwiftTokyoApp: App {
     }
   }
 }
+#endif
 
-struct ContentView: View {
+public struct ContentView: View {
   @State private var selectedTab = 0
 
-  var body: some View {
+  public init() {}
+
+  public var body: some View {
     TabView(selection: $selectedTab) {
       ScheduleScreen()
         .tabItem {
